+++ Scripts/Compare.cs
@@ -1,15 +1,19 @@
+using System;
+using System.Collections.Generic;
+using JetBrains.Annotations;
 using UnityEngine;
 
 namespace AeLa.Utilities
 {
-	public static class Comparers
+	[PublicAPI]
+	public static class Compare
 	{
 		/// <summary>
 		/// Can be used with a comparer in a Sort call to sort in descending order
 		/// </summary>
 		/// <example><code>positions.Sort((a, b) => Descend(Distance(a, b, origin)))</code></example>
 		public static int Descend(int comparison) => -comparison;
-		
+
 		/// <summary>
 		/// Compares the distance between a and b vectors and origin
 		/// </summary>
@@ -30,5 +34,15 @@ namespace AeLa.Utilities
 		/// <inheritdoc cref="Distance(UnityEngine.Vector3,UnityEngine.Vector3,UnityEngine.Vector3)"/>
 		public static int Distance(Component a, Component b, Vector3 origin) =>
 			Distance(a.gameObject, b.gameObject, origin);
+
+		public static T Min<T>(T a, T b) where T : IComparable
+		{
+			return Comparer<T>.Default.Compare(a, b) <= 0 ? a : b;
+		}
+
+		public static T Max<T>(T a, T b) where T : IComparable
+		{
+			return Comparer<T>.Default.Compare(a, b) <= 0 ? b : a;
+		}
 	}
 }