+++ Scripts/UI/TargetGraphicsGroup.cs
@@ -1,0 +1,11 @@
+using UnityEngine;
+using UnityEngine.UI;
+
+namespace AeLa.Utilities.UI
+{
+	// source: https://discussions.unity.com/t/tint-multiple-targets-with-single-button/556754/11
+	public class TargetGraphicsGroup : MonoBehaviour
+	{
+		public Graphic[] TargetGraphics;
+	}
+}